requirements.txt
# general  modules 
django
Pillow
djangorestframework

# env controlls
python-decouple

# third party modules
django-filter
django-cors-headers
drf_yasg
djoser
djangorestframework_simplejwt
social-auth-app-django

# deployment module
# gunicorn


# refactoring and tests
# black
# flake8
# pytest
# pytest-django
# Faker


# background proccess and cache
# celery
# django-redis
# Redis

# pygments





